requirements.txt
Flask==2.3.2
rembg==2.0.50
pillow==10.0.0